:recycle: Refactoring code. 租户ttl切换清空拦截器修改为过滤器

+++ pigx-common/pigx-common-data/src/main/java/com/pig4cloud/pigx/common/data/datascope/PigxBaseMapper.java
@@ -40,6 +40,6 @@ public interface PigxBaseMapper<T> extends BaseMapper<T> {
 	 * @param scope 数据权限范围
 	 * @return Integer
 	 */
-	Integer selectCountByScope(@Param(Constants.WRAPPER) Wrapper<T> queryWrapper, DataScope scope);
+	Long selectCountByScope(@Param(Constants.WRAPPER) Wrapper<T> queryWrapper, DataScope scope);
 
 }
